Add BufferedFanOut to project
A buffered fanout was an idea cooked in github.com/eloylp/meerkat project.
It provides an in memory fan out pattern to easily distribute data to
different parts of the project.In the case of meerkat , the motivation
was making a composition of the output of multiple public webcams on the
internet.Now is generalized with interface{} to allow us make us of it
with different types.