Change subscription structure in fanout
Before this change, internal management of subscribers were relaying in uuids. This makes deleting a subscriber an O(n) operation. Now we are relaying in the original array indexes internally, so this should be improved to O(1) or constant time on accesses. This also means we are leaving "gaps" in the internal subscriber array, which are being ignored. A future compaction mechanism could be added in the future.

Also, this uuids were returned to the user code whenever a new subscriber was added.  This makes impossible for users to properly choose a custom UUID format. After this change, we never return an UUID to the user, nor we allow any operation with them. We just allow providing a custom UUID that later the user can use to identify bottlenecks by calling the Status() function.