Make handler tester checkers use a copy of the body
A checker could try to consume the body buffer. If that happens, the
next checker will found a consumed Reader. In order to not propagate this body reconstruction logic across checkers, we just save a copy in the tester and create a reader per each checker, over the same underlying buffer.